car_picker/requirements.txt
streamlit>=1.36
pillow>=9.5
pyyaml>=6.0
